+++ tests/notify/test_templates.py
@@ -74,4 +74,4 @@ def test_gift_card_purchased_renders_amount_as_formatted_currency() -> None:
 
 def to_cents(dollars: float) -> int:
     """Dollars to integer cents (ADR-0002)."""
-    return int(round(dollars * 100))
+    return round(dollars * 100)
